backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
psycopg[binary]==3.3.2
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
openai==1.3.0
google-generativeai==0.3.1
google-auth==2.25.0
reportlab==4.0.7
openpyxl==3.1.2
